+++ goal_glide/commands/version_cmds.py
@@ -1,0 +1,15 @@
+from __future__ import annotations
+
+import click
+from rich.console import Console
+
+console = Console()
+
+
+@click.command("version")
+def version_cmds() -> None:
+    """Print package version."""
+    from .. import __version__
+
+    console.print(__version__)
+
