+++ tests/test_cli.py
@@ -4,7 +4,7 @@ import pytest
 import json
 from datetime import datetime, timedelta
 
-import goal_glide.cli as cli
+from goal_glide.cli import cli
 from goal_glide import config
 from goal_glide.models.storage import Storage
 from goal_glide.services import pomodoro
@@ -15,19 +15,18 @@ def test_add_list_remove(tmp_path):
 
     # add goal
     result = runner.invoke(
-        cli.goal, ["add", "Test Goal"], env={"GOAL_GLIDE_DB_DIR": str(tmp_path)}
+        cli, ["goal", "add", "Test Goal"], env={"GOAL_GLIDE_DB_DIR": str(tmp_path)}
     )
     assert result.exit_code == 0
 
     # list
-    result = runner.invoke(cli.goal, ["list"], env={"GOAL_GLIDE_DB_DIR": str(tmp_path)})
+    result = runner.invoke(cli, ["goal", "list"], env={"GOAL_GLIDE_DB_DIR": str(tmp_path)})
     assert "Test" in result.output
 
     # remove using id from storage (rich table may truncate id)
     goal_id = Storage(tmp_path).list_goals()[0].id
     result = runner.invoke(
-        cli.goal,
-        ["remove", goal_id],
+        cli, ["goal", "remove", goal_id],
         input="y\n",
         env={"GOAL_GLIDE_DB_DIR": str(tmp_path)},
     )
@@ -39,8 +38,8 @@ def test_add_duplicate_goal_warning(tmp_path) -> None:
     env = {"GOAL_GLIDE_DB_DIR": str(tmp_path)}
     runner = CliRunner()
 
-    runner.invoke(cli.goal, ["add", "dup"], env=env)
-    result = runner.invoke(cli.goal, ["add", "dup"], env=env)
+    runner.invoke(cli, ["goal", "add", "dup"], env=env)
+    result = runner.invoke(cli, ["goal", "add", "dup"], env=env)
 
     assert result.exit_code == 0
     assert "Warning: goal with this title already exists." in result.output
@@ -54,25 +53,22 @@ def test_pomo_session_persisted(tmp_path, monkeypatch):
     importlib.reload(pomodoro)
     runner = CliRunner()
     res = runner.invoke(
-        cli.goal, ["add", "G"], env={"GOAL_GLIDE_DB_DIR": str(tmp_path)}
+        cli, ["goal", "add", "G"], env={"GOAL_GLIDE_DB_DIR": str(tmp_path)}
     )
     gid = res.output.split()[-1].strip("()")
     runner.invoke(
-        cli.goal,
-        ["pomo", "start", "--duration", "1", "--goal", gid],
+        cli, ["goal", "pomo", "start", "--duration", "1", "--goal", gid],
         env={"GOAL_GLIDE_DB_DIR": str(tmp_path)},
     )
-    runner.invoke(cli.goal, ["pomo", "stop"], env={"GOAL_GLIDE_DB_DIR": str(tmp_path)})
+    runner.invoke(cli, ["goal", "pomo", "stop"], env={"GOAL_GLIDE_DB_DIR": str(tmp_path)})
     # ensure further pomodoro commands see no active session
     status = runner.invoke(
-        cli.goal,
-        ["pomo", "status"],
+        cli, ["goal", "pomo", "status"],
         env={"GOAL_GLIDE_DB_DIR": str(tmp_path)},
     )
     assert "No active session" in status.output
     paused = runner.invoke(
-        cli.goal,
-        ["pomo", "pause"],
+        cli, ["goal", "pomo", "pause"],
         env={"GOAL_GLIDE_DB_DIR": str(tmp_path)},
     )
     assert paused.exit_code == 1
@@ -89,11 +85,11 @@ def test_pomo_pause_resume(tmp_path, monkeypatch):
     import importlib
     importlib.reload(pomodoro)
     runner = CliRunner()
-    runner.invoke(cli.goal, ["pomo", "start", "--duration", "1"])
-    res = runner.invoke(cli.goal, ["pomo", "pause"])
+    runner.invoke(cli, ["goal", "pomo", "start", "--duration", "1"])
+    res = runner.invoke(cli, ["goal", "pomo", "pause"])
     assert res.exit_code == 0
     assert "paused" in res.output.lower()
-    res = runner.invoke(cli.goal, ["pomo", "resume"])
+    res = runner.invoke(cli, ["goal", "pomo", "resume"])
     assert res.exit_code == 0
     assert "resumed" in res.output.lower()
 
@@ -102,7 +98,7 @@ def test_jot_from_editor(tmp_path, monkeypatch):
     monkeypatch.setenv("GOAL_GLIDE_DB_DIR", str(tmp_path))
     monkeypatch.setattr(click, "edit", lambda *a, **k: "note from editor\n")
     runner = CliRunner()
-    result = runner.invoke(cli.thought, ["jot"])
+    result = runner.invoke(cli, ["thought", "jot"])
     assert result.exit_code == 0
     thought_text = Storage(tmp_path).list_thoughts()[0].text
     assert thought_text == "note from editor"
@@ -115,9 +111,9 @@ def test_goal_commands_invalid_id(cmd, goal_id, tmp_path):
     runner = CliRunner()
     args = [cmd, goal_id]
     if cmd == "remove":
-        result = runner.invoke(cli.goal, args, input="y\n", env=env)
+        result = runner.invoke(cli, ["goal", cmd, goal_id], input="y\n", env=env)
     else:
-        result = runner.invoke(cli.goal, args, env=env)
+        result = runner.invoke(cli, ["goal", cmd, goal_id], env=env)
     assert result.exit_code == 1
     assert "Error:" in result.output
 
@@ -126,7 +122,7 @@ def test_jot_from_editor_unicode(tmp_path, monkeypatch):
     monkeypatch.setenv("GOAL_GLIDE_DB_DIR", str(tmp_path))
     monkeypatch.setattr(click, "edit", lambda *a, **k: "Привет мир\n")
     runner = CliRunner()
-    result = runner.invoke(cli.thought, ["jot"])
+    result = runner.invoke(cli, ["thought", "jot"])
     assert result.exit_code == 0
     stored = Storage(tmp_path).list_thoughts()[0].text
     assert stored == "Привет мир"
@@ -136,7 +132,7 @@ def test_jot_from_editor_empty(tmp_path, monkeypatch):
     monkeypatch.setenv("GOAL_GLIDE_DB_DIR", str(tmp_path))
     monkeypatch.setattr(click, "edit", lambda *a, **k: "")
     runner = CliRunner()
-    result = runner.invoke(cli.thought, ["jot"])
+    result = runner.invoke(cli, ["thought", "jot"])
     assert result.exit_code == 1
     assert "Error:" in result.output
     assert Storage(tmp_path).list_thoughts() == []
@@ -145,7 +141,7 @@ def test_jot_from_editor_empty(tmp_path, monkeypatch):
 def test_config_quotes_disable(tmp_path, monkeypatch):
     monkeypatch.setattr(config, "_CONFIG_PATH", tmp_path / "config.toml")
     runner = CliRunner()
-    result = runner.invoke(cli.goal, ["config", "quotes", "--disable"])
+    result = runner.invoke(cli, ["goal", "config", "quotes", "--disable"])
     assert result.exit_code == 0
     assert "Quotes are OFF" in result.output
     assert config.quotes_enabled() is False
@@ -154,8 +150,8 @@ def test_config_quotes_disable(tmp_path, monkeypatch):
 def test_config_quotes_enable(tmp_path, monkeypatch):
     monkeypatch.setattr(config, "_CONFIG_PATH", tmp_path / "config.toml")
     runner = CliRunner()
-    runner.invoke(cli.goal, ["config", "quotes", "--disable"])
-    result = runner.invoke(cli.goal, ["config", "quotes", "--enable"])
+    runner.invoke(cli, ["goal", "config", "quotes", "--disable"])
+    result = runner.invoke(cli, ["goal", "config", "quotes", "--enable"])
     assert result.exit_code == 0
     assert "Quotes are ON" in result.output
     assert config.quotes_enabled() is True
@@ -169,15 +165,13 @@ def test_pomo_start_after_archive(tmp_path, monkeypatch):
     importlib.reload(pomodoro)
     runner = CliRunner()
     add_res = runner.invoke(
-        cli.goal,
-        ["add", "g"],
+        cli, ["goal", "add", "g"],
         env={"GOAL_GLIDE_DB_DIR": str(tmp_path)},
     )
     gid = add_res.output.split()[-1].strip("()")
-    runner.invoke(cli.goal, ["archive", gid], env={"GOAL_GLIDE_DB_DIR": str(tmp_path)})
+    runner.invoke(cli, ["goal", "archive", gid], env={"GOAL_GLIDE_DB_DIR": str(tmp_path)})
     start = runner.invoke(
-        cli.goal,
-        ["pomo", "start", "--duration", "1", "--goal", gid],
+        cli, ["goal", "pomo", "start", "--duration", "1", "--goal", gid],
         env={"GOAL_GLIDE_DB_DIR": str(tmp_path)},
     )
     assert start.exit_code == 0
@@ -193,8 +187,7 @@ def test_pomo_start_default_from_config(tmp_path, monkeypatch):
     monkeypatch.setattr(config, "pomo_duration", lambda: 2)
     runner = CliRunner()
     result = runner.invoke(
-        cli.goal,
-        ["pomo", "start"],
+        cli, ["goal", "pomo", "start"],
         env={"GOAL_GLIDE_DB_DIR": str(tmp_path)},
     )
     assert result.exit_code == 0
@@ -208,16 +201,16 @@ def test_list_due_filters(tmp_path):
     soon = (datetime.utcnow() + timedelta(days=2)).strftime("%Y-%m-%d")
     later = (datetime.utcnow() + timedelta(days=5)).strftime("%Y-%m-%d")
     past = (datetime.utcnow() - timedelta(days=1)).strftime("%Y-%m-%d")
-    runner.invoke(cli.goal, ["add", "soon", "--deadline", soon], env=env)
-    runner.invoke(cli.goal, ["add", "later", "--deadline", later], env=env)
-    runner.invoke(cli.goal, ["add", "past", "--deadline", past], env=env)
+    runner.invoke(cli, ["goal", "add", "soon", "--deadline", soon], env=env)
+    runner.invoke(cli, ["goal", "add", "later", "--deadline", later], env=env)
+    runner.invoke(cli, ["goal", "add", "past", "--deadline", past], env=env)
 
-    res = runner.invoke(cli.goal, ["list", "--due-soon"], env=env)
+    res = runner.invoke(cli, ["goal", "list", "--due-soon"], env=env)
     assert "soon" in res.output
     assert "later" not in res.output
     assert "past" not in res.output
 
-    res = runner.invoke(cli.goal, ["list", "--overdue"], env=env)
+    res = runner.invoke(cli, ["goal", "list", "--overdue"], env=env)
     assert "past" in res.output
     assert "soon" not in res.output
     assert "later" not in res.output
