+++ tests/test_goal_subgoals.py
@@ -4,7 +4,7 @@ import tempfile
 
 from click.testing import CliRunner
 
-from goal_glide.cli import goal
+from goal_glide.cli import cli
 from goal_glide.models.goal import Goal
 from goal_glide.models.storage import Storage
 from hypothesis import given, settings, strategies as st
@@ -121,9 +121,9 @@ def test_list_goals_parent_property(mapping: dict[str, list[str]]) -> None:
 def test_cli_add_with_parent(tmp_path: Path) -> None:
     runner = CliRunner()
     env = {"GOAL_GLIDE_DB_DIR": str(tmp_path)}
-    runner.invoke(goal, ["add", "parent"], env=env)
+    runner.invoke(cli, ["goal", "add", "parent"], env=env)
     pid = Storage(tmp_path).list_goals()[0].id
-    runner.invoke(goal, ["add", "child", "--parent", pid], env=env)
+    runner.invoke(cli, ["goal", "add", "child", "--parent", pid], env=env)
     children = Storage(tmp_path).list_goals(parent_id=pid)
     assert len(children) == 1 and children[0].title == "child"
 
@@ -131,9 +131,9 @@ def test_cli_add_with_parent(tmp_path: Path) -> None:
 def test_goal_tree_output(tmp_path: Path) -> None:
     runner = CliRunner()
     env = {"GOAL_GLIDE_DB_DIR": str(tmp_path)}
-    runner.invoke(goal, ["add", "parent"], env=env)
+    runner.invoke(cli, ["goal", "add", "parent"], env=env)
     pid = Storage(tmp_path).list_goals()[0].id
-    runner.invoke(goal, ["add", "child", "--parent", pid], env=env)
-    result = runner.invoke(goal, ["tree"], env=env)
+    runner.invoke(cli, ["goal", "add", "child", "--parent", pid], env=env)
+    result = runner.invoke(cli, ["goal", "tree"], env=env)
     assert "child" in result.output
     assert result.output.find("child") > result.output.find("parent")
