+++ tests/test_pomo_quotes.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import pytest
 from click.testing import CliRunner
 
-from goal_glide import cli
+from goal_glide.cli import cli
 from goal_glide import config as cfg
 from goal_glide.services import quotes
 
@@ -23,8 +23,8 @@ def test_pomo_stop_prints_quote(
 ) -> None:
     monkeypatch.setattr(quotes, "get_random_quote", lambda use_online=True: ("Q", "A"))
     monkeypatch.setattr(cli, "get_random_quote", lambda use_online=True: ("Q", "A"))
-    runner.invoke(cli.goal, ["pomo", "start", "--duration", "1"])
-    result = runner.invoke(cli.goal, ["pomo", "stop"])
+    runner.invoke(cli, ["pomo", "start", "--duration", "1"])
+    result = runner.invoke(cli, ["pomo", "stop"])
     assert "Pomodoro complete" in result.output
     assert "Q" in result.output
 
@@ -37,8 +37,8 @@ def test_quotes_disabled(
     path.write_text("quotes_enabled = false", encoding="utf-8")
     monkeypatch.setattr(quotes, "get_random_quote", lambda use_online=True: ("Q", "A"))
     monkeypatch.setattr(cli, "get_random_quote", lambda use_online=True: ("Q", "A"))
-    runner.invoke(cli.goal, ["pomo", "start", "--duration", "1"])
-    result = runner.invoke(cli.goal, ["pomo", "stop"])
+    runner.invoke(cli, ["pomo", "start", "--duration", "1"])
+    result = runner.invoke(cli, ["pomo", "stop"])
     assert "Pomodoro complete" in result.output
     assert "Q" not in result.output
 
@@ -52,8 +52,8 @@ def test_quote_fallback(monkeypatch: pytest.MonkeyPatch, runner: CliRunner) -> N
     monkeypatch.setattr(quotes, "_LOCAL_CACHE", None)
     monkeypatch.setattr(quotes.random, "choice", lambda seq: seq[0])
     monkeypatch.setattr(cli, "get_random_quote", quotes.get_random_quote)
-    runner.invoke(cli.goal, ["pomo", "start", "--duration", "1"])
-    result = runner.invoke(cli.goal, ["pomo", "stop"])
+    runner.invoke(cli, ["pomo", "start", "--duration", "1"])
+    result = runner.invoke(cli, ["pomo", "stop"])
     assert "Inspirational quote 1" in result.output
 
 
@@ -65,8 +65,8 @@ def test_quote_exception_handling(
 
     monkeypatch.setattr(quotes, "get_random_quote", boom)
     monkeypatch.setattr(cli, "get_random_quote", boom)
-    runner.invoke(cli.goal, ["pomo", "start", "--duration", "1"])
-    result = runner.invoke(cli.goal, ["pomo", "stop"])
+    runner.invoke(cli, ["pomo", "start", "--duration", "1"])
+    result = runner.invoke(cli, ["pomo", "stop"])
     assert result.exit_code == 1
     assert "unexpected" in result.output.lower()
 
@@ -74,7 +74,7 @@ def test_quote_exception_handling(
 def test_quotes_default_enabled(
     monkeypatch: pytest.MonkeyPatch, runner: CliRunner
 ) -> None:
-    result = runner.invoke(cli.goal, ["config", "quotes"])
+    result = runner.invoke(cli, ["config", "quotes"])
     assert result.exit_code == 0
     assert "Quotes are ON" in result.output
 
@@ -93,8 +93,8 @@ def test_quotes_disabled_no_call(
 
     monkeypatch.setattr(quotes, "get_random_quote", fake)
     monkeypatch.setattr(cli, "get_random_quote", fake)
-    runner.invoke(cli.goal, ["pomo", "start", "--duration", "1"])
-    result = runner.invoke(cli.goal, ["pomo", "stop"])
+    runner.invoke(cli, ["pomo", "start", "--duration", "1"])
+    result = runner.invoke(cli, ["pomo", "stop"])
     assert called == []
     assert "Pomodoro complete" in result.output
     assert "Q" not in result.output
