+++ tests/test_pomo_status.py
@@ -3,7 +3,7 @@ from pathlib import Path
 
 from click.testing import CliRunner
 
-from goal_glide import cli
+from goal_glide.cli import cli
 from goal_glide.services import pomodoro
 
 
@@ -14,7 +14,7 @@ def test_status_no_session(tmp_path: Path, monkeypatch):
     import importlib
     importlib.reload(pomodoro)
     runner = CliRunner()
-    result = runner.invoke(cli.goal, ["pomo", "status"])
+    result = runner.invoke(cli, ["goal", "pomo", "status"])
     assert result.exit_code == 0
     assert "No active session" in result.output
 
@@ -44,7 +44,7 @@ def test_status_with_session(tmp_path: Path, monkeypatch):
 
     monkeypatch.setattr(cli, "datetime", LaterDT)
     runner = CliRunner()
-    result = runner.invoke(cli.goal, ["pomo", "status"])
+    result = runner.invoke(cli, ["goal", "pomo", "status"])
     assert result.exit_code == 0
     assert "Elapsed 10m" in result.output
     assert "Remaining 20m" in result.output
@@ -80,7 +80,7 @@ def test_status_paused(tmp_path: Path, monkeypatch):
 
     monkeypatch.setattr(cli, "datetime", LaterDT)
     runner = CliRunner()
-    result = runner.invoke(cli.goal, ["pomo", "status"])
+    result = runner.invoke(cli, ["goal", "pomo", "status"])
     assert result.exit_code == 0
     assert "Elapsed 10m" in result.output
     assert "Remaining 20m" in result.output
