+++ tests/test_reminder.py
@@ -9,7 +9,7 @@ import types
 import pytest
 from click.testing import CliRunner
 
-from goal_glide import cli
+from goal_glide.cli import cli
 from goal_glide import config as cfg
 from goal_glide.services import notify, reminder
 
@@ -23,28 +23,28 @@ def runner(monkeypatch: pytest.MonkeyPatch, tmp_path: Path) -> CliRunner:
 
 
 def test_enable_disable_updates_config(runner: CliRunner) -> None:
-    runner.invoke(cli.goal, ["reminder", "enable"])
+    runner.invoke(cli, ["reminder", "enable"])
     assert cfg.reminders_enabled() is True
-    runner.invoke(cli.goal, ["reminder", "disable"])
+    runner.invoke(cli, ["reminder", "disable"])
     assert cfg.reminders_enabled() is False
 
 
 def test_config_command_updates_values(runner: CliRunner) -> None:
-    runner.invoke(cli.goal, ["reminder", "config", "--break", "10", "--interval", "15"])
+    runner.invoke(cli, ["reminder", "config", "--break", "10", "--interval", "15"])
     assert cfg.reminder_break() == 10
     assert cfg.reminder_interval() == 15
 
 
 @pytest.mark.parametrize("val", [0, -5, 200])
 def test_invalid_break_value_errors(val: int, runner: CliRunner) -> None:
-    result = runner.invoke(cli.goal, ["reminder", "config", "--break", str(val)])
+    result = runner.invoke(cli, ["reminder", "config", "--break", str(val)])
     assert result.exit_code != 0
     assert "break must be between 1 and 120" in result.output
 
 
 @pytest.mark.parametrize("val", [0, -5, 200])
 def test_invalid_interval_value_errors(val: int, runner: CliRunner) -> None:
-    result = runner.invoke(cli.goal, ["reminder", "config", "--interval", str(val)])
+    result = runner.invoke(cli, ["reminder", "config", "--interval", str(val)])
     assert result.exit_code != 0
     assert "interval must be between 1 and 120" in result.output
 
@@ -113,16 +113,16 @@ def test_schedule_after_stop_creates_scheduler(monkeypatch: pytest.MonkeyPatch) 
 
 
 def test_reminder_status_output(runner: CliRunner) -> None:
-    result = runner.invoke(cli.goal, ["reminder", "status"])
+    result = runner.invoke(cli, ["reminder", "status"])
     assert result.exit_code == 0
     assert "Enabled: False | Break: 5m | Interval: 30m" in result.output
 
-    runner.invoke(cli.goal, ["reminder", "enable"])
+    runner.invoke(cli, ["reminder", "enable"])
     runner.invoke(
-        cli.goal,
+        cli,
         ["reminder", "config", "--break", "11", "--interval", "22"],
     )
-    result = runner.invoke(cli.goal, ["reminder", "status"])
+    result = runner.invoke(cli, ["reminder", "status"])
     assert result.exit_code == 0
     assert "Enabled: True | Break: 11m | Interval: 22m" in result.output
