+++ tests/test_stats_cmd.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import pytest
 from click.testing import CliRunner
 
-from goal_glide import cli
+from goal_glide.cli import cli
 from goal_glide.models.session import PomodoroSession
 from goal_glide.models.storage import Storage
 
@@ -41,7 +41,7 @@ def test_stats_week_output_has_7_bars(
 
     monkeypatch.setattr(cli, "datetime", FakeDT)
     result = runner.invoke(
-        cli.goal, ["stats"], env={"GOAL_GLIDE_DB_DIR": str(tmp_path)}
+        cli, ["stats"], env={"GOAL_GLIDE_DB_DIR": str(tmp_path)}
     )
     days = ["Mon", "Tue", "Wed", "Thu", "Fri", "Sat", "Sun"]
     lines = [
@@ -69,7 +69,7 @@ def test_stats_month_output_has_4_bars(
 
     monkeypatch.setattr(cli, "datetime", FakeDT)
     result = runner.invoke(
-        cli.goal, ["stats", "--month"], env={"GOAL_GLIDE_DB_DIR": str(tmp_path)}
+        cli, ["stats", "--month"], env={"GOAL_GLIDE_DB_DIR": str(tmp_path)}
     )
     weeks = ["W1", "W2", "W3", "W4"]
     lines = [
@@ -96,7 +96,7 @@ def test_stats_goals_table_shows_top5(
 
     monkeypatch.setattr(cli, "datetime", FakeDT)
     result = runner.invoke(
-        cli.goal, ["stats", "--goals"], env={"GOAL_GLIDE_DB_DIR": str(tmp_path)}
+        cli, ["stats", "--goals"], env={"GOAL_GLIDE_DB_DIR": str(tmp_path)}
     )
     assert result.exit_code == 0
     assert "Top Goals" in result.output
@@ -114,7 +114,7 @@ def test_stats_empty_db_graceful(
 
     monkeypatch.setattr(cli, "datetime", FakeDT)
     result = runner.invoke(
-        cli.goal, ["stats"], env={"GOAL_GLIDE_DB_DIR": str(tmp_path)}
+        cli, ["stats"], env={"GOAL_GLIDE_DB_DIR": str(tmp_path)}
     )
     assert result.exit_code == 0
     assert "No session data" in result.output
@@ -135,7 +135,7 @@ def test_stats_custom_range(
 
     monkeypatch.setattr(cli, "datetime", FakeDT)
     result = runner.invoke(
-        cli.goal,
+        cli,
         ["stats", "--from", "2023-01-01", "--to", "2023-01-03"],
         env={"GOAL_GLIDE_DB_DIR": str(tmp_path)},
     )
