+++ tests/test_thoughts.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import pytest
 from click.testing import CliRunner
 
-from goal_glide.cli import goal, thought
+from goal_glide.cli import cli
 from goal_glide.models.storage import Storage
 from goal_glide.models.thought import Thought
 
@@ -18,7 +18,7 @@ def runner(monkeypatch, tmp_path: Path) -> CliRunner:
 
 
 def test_jot_basic(tmp_path: Path, runner: CliRunner) -> None:
-    result = runner.invoke(thought, ["jot", "note"])
+    result = runner.invoke(cli, ["thought", "jot", "note"])
     assert result.exit_code == 0
     storage = Storage(tmp_path)
     thoughts = storage.list_thoughts()
@@ -28,16 +28,16 @@ def test_jot_basic(tmp_path: Path, runner: CliRunner) -> None:
 
 
 def test_jot_with_goal(tmp_path: Path, runner: CliRunner) -> None:
-    runner.invoke(goal, ["add", "goal 1"])
+    runner.invoke(cli, ["goal", "add", "goal 1"])
     goal_id = Storage(tmp_path).list_goals()[0].id
-    result = runner.invoke(thought, ["jot", "idea", "-g", goal_id])
+    result = runner.invoke(cli, ["thought", "jot", "idea", "-g", goal_id])
     assert result.exit_code == 0
     t = Storage(tmp_path).list_thoughts()[0]
     assert t.goal_id == goal_id
 
 
 def test_jot_blank_fails(tmp_path: Path, runner: CliRunner) -> None:
-    result = runner.invoke(thought, ["jot", "  "])
+    result = runner.invoke(cli, ["thought", "jot", "  "])
     assert result.exit_code != 0
     assert not Storage(tmp_path).list_thoughts()
 
@@ -48,7 +48,7 @@ def test_list_default_order(tmp_path: Path, runner: CliRunner) -> None:
     newer = Thought(id="2", text="new", timestamp=datetime.now())
     storage.add_thought(older)
     storage.add_thought(newer)
-    result = runner.invoke(thought, ["list"])
+    result = runner.invoke(cli, ["thought", "list"])
     rows = [line for line in result.output.splitlines() if "│" in line]
     assert "new" in rows[0]
     assert "old" in rows[1]
@@ -58,20 +58,20 @@ def test_list_limit(tmp_path: Path, runner: CliRunner) -> None:
     storage = Storage(tmp_path)
     for i in range(5):
         storage.add_thought(Thought(id=str(i), text=f"t{i}", timestamp=datetime.now()))
-    result = runner.invoke(thought, ["list", "--limit", "3"])
+    result = runner.invoke(cli, ["thought", "list", "--limit", "3"])
     assert result.exit_code == 0
     rows = [line for line in result.output.splitlines() if "│" in line]
     assert len(rows) <= 3
 
 
 def test_list_goal_filter(tmp_path: Path, runner: CliRunner) -> None:
-    runner.invoke(goal, ["add", "g"])
+    runner.invoke(cli, ["goal", "add", "g"])
     goal_id = Storage(tmp_path).list_goals()[0].id
     Storage(tmp_path).add_thought(Thought(id="1", text="a", timestamp=datetime.now()))
     Storage(tmp_path).add_thought(
         Thought(id="2", text="b", timestamp=datetime.now(), goal_id=goal_id)
     )
-    result = runner.invoke(thought, ["list", "-g", goal_id])
+    result = runner.invoke(cli, ["thought", "list", "-g", goal_id])
     rows = [line for line in result.output.splitlines() if "│" in line]
     assert any("b" in r for r in rows)
     assert all(r.split("│")[3].strip() != "a" for r in rows)
@@ -91,12 +91,12 @@ def test_migration_keeps_other_tables(tmp_path: Path, runner: CliRunner) -> None
 def test_remove_thought(tmp_path: Path, runner: CliRunner) -> None:
     t = Thought(id="x", text="bye", timestamp=datetime.now())
     Storage(tmp_path).add_thought(t)
-    result = runner.invoke(thought, ["rm", "x"])
+    result = runner.invoke(cli, ["thought", "rm", "x"])
     assert result.exit_code == 0
     assert not Storage(tmp_path).list_thoughts()
 
 
 def test_remove_thought_missing(tmp_path: Path, runner: CliRunner) -> None:
-    result = runner.invoke(thought, ["rm", "bad"])
+    result = runner.invoke(cli, ["thought", "rm", "bad"])
     assert result.exit_code == 0
     assert "not found" in result.output
